+++ src/main/dll/DR/dll_0253_ktlazerlight.c
@@ -72,7 +72,7 @@ void ktlazerlight_update(int obj)
     char* extra = ((GameObject*)obj)->extra;
     s16 intensity;
     void* light = *(void**)(extra + 0x4);
-    intensity = (s16)GameBit_Get(((KtlazerlightPlacement*)placement)->onIntensityBit);
+    intensity = GameBit_Get(((KtlazerlightPlacement*)placement)->onIntensityBit);
     if (intensity >= 1 || GameBit_Get(((KtlazerlightPlacement*)placement)->onStayLitBit) != 0)
     {
         if (intensity == 0)
