+++ src/main/dll/dll_009B_dll9bfunc0.c
@@ -139,7 +139,7 @@ void dll_9B_func03(int target, int variant, int parent, u32 flags)
 
     hdr.v0 = 0;
     hdr.target = target;
-    hdr.b = (s16)variant;
+    hdr.b = variant;
     hdr.bx = lbl_803E13A0;
     hdr.by = lbl_803E13A0;
     hdr.bz = lbl_803E13A0;
